+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### Changed
+
+- Get all likes for a particular listing or profile will show the necessary profile and listing information respectively.
+
 ## [0.0.3] - 2020-08-13
 
 ### Added